Make toDiffEqual globally accessible.
